requirements.txt
# Discord
discord.py==2.3.2

# HTTP Requests
aiohttp==3.9.1
requests==2.31.0

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Database
aiosqlite==0.19.0

# Utilities
python-dateutil==2.8.2

curl-cffi>=0.5.0